Move the test helper for plugin to a seperate go file
